+++ src/controllers/game.controller.js
@@ -2,7 +2,7 @@ const httpStatus = require('http-status');
 const pick = require('../utils/pick');
 const ApiError = require('../utils/ApiError');
 const catchAsync = require('../utils/catchAsync');
-const { gameService } = require('../services');
+const { gameService, messageService, memberService, betService, tipService, loggingService } = require('../services');
 const bannerUrls = require('../utils/bannerUrls');
 
 const createGame = catchAsync(async (req, res) => {
@@ -16,7 +16,7 @@ const createGame = catchAsync(async (req, res) => {
   const game = await gameService.createGame(gameBody);
 
   const logTitle = game.title.length > 50 ? game.title.sustr(0,48) + '..' : game.title;
-  await gameService.addLog(game.id,'gameCreated','Game created','Game ' + logTitle + ' has been created.');
+  await loggingService.createLogging({gameId: game.id, logType: 'gameCreated', title: 'Game created', desc: 'Game ' + logTitle + ' has been created.'});
 
   res.status(httpStatus.CREATED).send(game);
 });
@@ -55,17 +55,32 @@ const updateGame = catchAsync(async (req, res) => {
   res.send(game);
 });
 
-/*
 const deleteGame = catchAsync(async (req, res) => {
+  if (!req.user) 
+    throw new ApiError(httpStatus.UNAUTHORIZED, 'Not authorized.');
+
+  let game = await gameService.getGameById(req.params.gameId);
+  if (!game) 
+    throw new ApiError(httpStatus.NOT_FOUND, 'Game not found.');
+
+  if (game.userId != req.user.id)
+    throw new ApiError(httpStatus.FORBIDDEN, 'Not authorized to delete this game.');
+  
+  await tipService.deleteTipsByGameId(req.params.gameId);
+  await messageService.deleteMessagesByGameId(req.params.gameId);
+  await memberService.deleteMembersByGameId(req.params.gameId);
+  await betService.deleteBetsByGameId(req.params.gameId);
+  await loggingService.deleteLoggingsByGameId(req.params.gameId);
+    
   await gameService.deleteGameById(req.params.gameId);
   res.status(httpStatus.NO_CONTENT).send();
 });
-*/
+
 
 module.exports = {
   createGame,
   getGames,
   getGame,
   updateGame,
-  //deleteGame,
+  deleteGame,
 };
